retriever/requirements.txt
fastapi
pydantic
faiss-cpu==1.7.4
sentence-transformers
numpy<2.0
uvicorn
